requirements.txt
pytesseract
Pillow
pandas
python-dotenv
groq
logging
pathlib
streamlit